Some code resturcturing
pulled completed functions out into seperate .h and .cpp files
