+++ SliderAI.exe.xcodeproj/xcuserdata/aaron.xcuserdatad/xcschemes/SliderAI.xcscheme
@@ -17,7 +17,7 @@
                BlueprintIdentifier = "DC41E49F1E5C03BC0080DFB3"
                BuildableName = "SliderAI"
                BlueprintName = "SliderAI"
-               ReferencedContainer = "container:SliderAI.xcodeproj">
+               ReferencedContainer = "container:SliderAI.exe.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -35,7 +35,7 @@
             BlueprintIdentifier = "DC41E49F1E5C03BC0080DFB3"
             BuildableName = "SliderAI"
             BlueprintName = "SliderAI"
-            ReferencedContainer = "container:SliderAI.xcodeproj">
+            ReferencedContainer = "container:SliderAI.exe.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -58,7 +58,7 @@
             BlueprintIdentifier = "DC41E49F1E5C03BC0080DFB3"
             BuildableName = "SliderAI"
             BlueprintName = "SliderAI"
-            ReferencedContainer = "container:SliderAI.xcodeproj">
+            ReferencedContainer = "container:SliderAI.exe.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -77,7 +77,7 @@
             BlueprintIdentifier = "DC41E49F1E5C03BC0080DFB3"
             BuildableName = "SliderAI"
             BlueprintName = "SliderAI"
-            ReferencedContainer = "container:SliderAI.xcodeproj">
+            ReferencedContainer = "container:SliderAI.exe.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
