+++ SliderAI.exe.xcodeproj/xcuserdata/stephen.xcuserdatad/xcschemes/xcschememanagement.plist
