--- SliderAI/sliderai.cpp
@@ -1,8 +1,0 @@
-//
-//  sliderai.cpp
-//  SliderAI
-//
-
-#include "sliderai.h"
-
-
